renamed folder and files to indicate that those aren't components
