+++ client/components/Bomb/modules/clock.js
