+++ client/components/Bomb/modules/symbols.js
