+++ client/components/Bomb/modules/wires.js
